Restore nvim-spider operator pending bindings
This fixes and restores the use of nvim-spider's operator pending
bindings. I'd previously removed those for the w,e,b motions as too
intrusive, turning "do a thing to the word" into "do a thing to <count>
camel-words". I fixed this by a separation of bindings - w,e,b are used
for plain motions, and ",w", ",e", ",b" are used for operator pending
mode ala the old vim CamelCaseMotion plugin.

Also fixed the broken use of `lazy` in this plugin's config. Now it uses
the VeryLazy event so that it's always configured.